services/rag-indexer/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.49.0
google-cloud-storage==2.18.2
numpy==1.26.4
